pyproject.toml
[tool.poetry]
name = "forest"
version = "0.1.0"
description = "Forest — AI-native virtual file system for Discord"
authors = []
readme = "README.md"
packages = [{ include = "forest" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = { extras = ["standard"], version = "^0.32.0" }
discord-py = "^2.4.0"
sqlalchemy = { extras = ["asyncio"], version = "^2.0.36" }
asyncpg = "^0.30.0"
alembic = "^1.14.0"
pydantic-settings = "^2.6.0"
openai = "^1.57.0"
pgvector = "^0.3.6"
httpx = "^0.28.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
ruff = "^0.8.0"
psycopg2-binary = "^2.9.10"

[tool.poetry.scripts]
forest = "forest.main:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.ruff.lint.per-file-ignores]
"alembic/**" = ["E501", "I001", "UP007", "UP035"]
